Factor out anki tab
